refactor: :recycle: move limit selection into own function
